+++ VKDesign(11September)/Utils/CoreDataManager.swift
@@ -1,0 +1,43 @@
+//
+//  CoreDataManager.swift
+//  VKDesign(11September)
+//
+//  Created by BLVCK on 18/12/2017.
+//  Copyright © 2017 blvvvck production. All rights reserved.
+//
+
+import Foundation
+import CoreData
+
+class CoreDataManager {
+    
+    static let instance = CoreDataManager()
+    
+    lazy var persistentContainer : NSPersistentContainer = {
+        
+        let container = NSPersistentContainer(name: "VKDesign(11September)")
+        
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+    
+    func saveContext() {
+        let context = persistentContainer.viewContext
+        
+        if context.hasChanges {
+            
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+        
+    }
+}
